Correct sources ignored by 'strict-dynamic'
Split protocol/host sources into two different lines for clarity

In the CSP3 draft spec (https://www.w3.org/TR/CSP3/#strict-dynamic-usage)
'unsafe-eval' is not present in the list of sources ignored when
'strict-dynamic' is present in the policy. Implementations by current
versions of Chromium and Firefox do not ignore 'unsafe-eval' with
'strict-dynamic'.